+++ dotcom-rendering/src/article.ts
@@ -1,20 +1,41 @@
 import type { ArticleFormat } from '@guardian/libs';
-import { appsLightboxImages } from '../model/appsLightboxImages';
-import { buildLightboxImages } from '../model/buildLightboxImages';
-import { enhanceBlocks, enhanceMainMedia } from '../model/enhanceBlocks';
-import { enhanceCommercialProperties } from '../model/enhanceCommercialProperties';
-import { enhanceStandfirst } from '../model/enhanceStandfirst';
-import { enhanceTableOfContents } from '../model/enhanceTableOfContents';
-import { validateAsArticleType } from '../model/validate';
-import { type DCRArticle } from '../types/frontend';
-import { type RenderingTarget } from '../types/renderingTarget';
-import { decideFormat } from './decideFormat';
+import { decideFormat } from './lib/decideFormat';
+import type { ImageForAppsLightbox } from './model/appsLightboxImages';
+import { appsLightboxImages } from './model/appsLightboxImages';
+import { buildLightboxImages } from './model/buildLightboxImages';
+import { enhanceBlocks, enhanceMainMedia } from './model/enhanceBlocks';
+import { enhanceCommercialProperties } from './model/enhanceCommercialProperties';
+import { enhanceStandfirst } from './model/enhanceStandfirst';
+import { enhanceTableOfContents } from './model/enhanceTableOfContents';
+import { validateAsArticleType } from './model/validate';
+import type { ImageForLightbox } from './types/content';
+import type { FEArticleType } from './types/frontend';
+import { type RenderingTarget } from './types/renderingTarget';
+
+/**
+ * The `DCRArticle` type models the `FEArticleType` in addition to any enhancements DCR makes after
+ * receiving the data from Frontend.
+ */
+export type Article = FEArticleType & {
+	imagesForLightbox: ImageForLightbox[];
+	imagesForAppsLightbox: ImageForAppsLightbox[];
+	tableOfContents?: TableOfContentsItem[];
+};
+
+export interface TableOfContents {
+	items: TableOfContentsItem[];
+}
+
+export interface TableOfContentsItem {
+	id: string;
+	title: string;
+}
 
 const enhancePinnedPost = (
 	format: ArticleFormat,
 	renderingTarget: RenderingTarget,
 	block?: Block,
-) => {
+): Block | undefined => {
 	if (!block) return;
 
 	return enhanceBlocks([block], format, {
@@ -28,7 +49,7 @@ const enhancePinnedPost = (
 export const enhanceArticleType = (
 	body: unknown,
 	renderingTarget: RenderingTarget,
-): DCRArticle => {
+): Article => {
 	const data = validateAsArticleType(body);
 	const format = decideFormat(data.format);
