+++ dotcom-rendering/src/components/ArticleBody.tsx
@@ -7,13 +7,13 @@ import {
 	headlineBold24,
 	remSpace,
 } from '@guardian/source/foundations';
+import type { TableOfContentsItem } from '../article';
 import { ArticleRenderer } from '../lib/ArticleRenderer';
 import type { EditionId } from '../lib/edition';
 import { decideLanguage, decideLanguageDirection } from '../lib/lang';
 import { revealStyles } from '../lib/revealStyles';
 import { palette as themePalette } from '../palette';
 import type { ServerSideTests, Switches } from '../types/config';
-import type { TableOfContentsItem } from '../types/frontend';
 import type { TagType } from '../types/tag';
 import { Island } from './Island';
 import { LiveBlogRenderer } from './LiveBlogRenderer';
