+++ dotcom-rendering/src/lib/contributions.ts
@@ -6,8 +6,8 @@ import {
 } from '@guardian/libs';
 import type { HeaderPayload } from '@guardian/support-dotcom-components/dist/dotcom/src/types';
 import { useEffect, useState } from 'react';
+import type { Article } from '../article';
 import type { DCRFrontType } from '../types/front';
-import type { DCRArticle } from '../types/frontend';
 import type { DCRNewslettersPageType } from '../types/newslettersPage';
 import type { DCRTagPageType } from '../types/tagPage';
 
@@ -269,7 +269,7 @@ export const recentlyClosedBanner = (
 };
 
 export const getContributionsServiceUrl = (
-	config: DCRArticle | DCRFrontType | DCRTagPageType | DCRNewslettersPageType,
+	config: Article | DCRFrontType | DCRTagPageType | DCRNewslettersPageType,
 ): string => process.env.SDC_URL ?? config.contributionsServiceUrl;
 
 type PurchaseInfo = HeaderPayload['targeting']['purchaseInfo'];
