+++ dotcom-rendering/src/lib/layoutHelpers.ts
@@ -1,8 +1,9 @@
 import type { Pillar } from '@guardian/libs';
-import type { DCRArticle, LegacyPillar } from '../types/frontend';
+import type { Article } from '../article';
+import type { LegacyPillar } from '../types/frontend';
 import { decideNavPillar } from './decideNavPillar';
 
-export const getCurrentPillar = (article: DCRArticle): Pillar | undefined => {
+export const getCurrentPillar = (article: Article): Pillar | undefined => {
 	const currentPillar =
 		(article.nav.currentPillarTitle?.toLowerCase() as
 			| LegacyPillar
