+++ dotcom-rendering/src/server/handler.article.web.ts
@@ -1,6 +1,6 @@
 import type { RequestHandler } from 'express';
 import { Standard as ExampleArticle } from '../../fixtures/generated/fe-articles/Standard';
-import { enhanceArticleType } from '../lib/article';
+import { enhanceArticleType } from '../article';
 import { decideFormat } from '../lib/decideFormat';
 import { enhanceBlocks } from '../model/enhanceBlocks';
 import { validateAsBlock } from '../model/validate';
